+++ src/datanode.hpp
@@ -63,7 +63,7 @@ struct DataNode {
     DataNode* AddArrayElemChild(const char* key, const DataNode& value);
 
     const char* Get(const char* key, const char* def="", bool quiet=false) const;
-    int GetI(const char* key, int def=0, bool quiet=false) const;
+    long GetI(const char* key, long def=0, bool quiet=false) const;
     double GetF(const char* key, double def=0, bool quiet=false) const;
     timemath::Time GetDate(const char* key, timemath::Time def=0, bool quiet=false) const;
     DataNode* GetChild(const char* key, bool quiet=false) const;
@@ -71,7 +71,7 @@ struct DataNode {
     size_t GetArrayLen(const char* key, bool quiet=false) const;
     size_t GetArrayChildLen(const char* key, bool quiet=false) const;
     const char* GetArray(const char* key, int index, const char* def = "", bool quiet=false) const;
-    int GetArrayI(const char* key, int index, int def=0, bool quiet=false) const;
+    long GetArrayI(const char* key, int index, long def=0, bool quiet=false) const;
     double GetArrayF(const char* key, int index, double def=0, bool quiet=false) const;
     DataNode* GetArrayChild(const char* key, int index, bool quiet=false) const;
